requirements.txt
pandas
numpy
SQLAlchemy
yfinance
python-dotenv
pyarrow
